고급편/섹션4: 프록시 패턴 - 예제 코드2 (#5)
섹션4: 프록시 패턴과 데코레이터 패턴
