--- 스프링 로드맵5 고급편/proxy/src/test/java/hello/proxy/proxy/ProxyPatternTest.java
@@ -1,17 +1,0 @@
-package hello.proxy.proxy;
-
-import hello.proxy.pureproxy.code.ProxyPatternClient;
-import hello.proxy.pureproxy.code.RealSubject;
-import org.junit.jupiter.api.Test;
-
-public class ProxyPatternTest {
-
-    @Test
-    void noProxyTest() {
-        RealSubject realSubject = new RealSubject();
-        ProxyPatternClient client = new ProxyPatternClient(realSubject);
-        client.execute();
-        client.execute();
-        client.execute();
-    }
-}
