+++ 스프링 로드맵5 고급편/proxy/src/test/java/hello/proxy/pureproxy/proxy/code/CacheProxy.java
@@ -1,0 +1,26 @@
+package hello.proxy.pureproxy.proxy.code;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class CacheProxy implements Subject {
+
+    private Subject target;
+    private String cacheValue;
+
+    public CacheProxy(final Subject target) {
+        this.target = target;
+    }
+
+    @Override
+    public String operation() {
+        log.info("프록시 호출");
+        if (cacheValue == null) {
+            // 첫 호출시에 실제 객체를 호출하여 Data 반환
+            cacheValue = target.operation();
+        }
+
+        return cacheValue;
+    }
+
+}
