+++ 스프링 로드맵5 고급편/proxy/src/test/java/hello/proxy/pureproxy/proxy/code/Subject.java
@@ -1,4 +1,4 @@
-package hello.proxy.pureproxy.code;
+package hello.proxy.pureproxy.proxy.code;
 
 public interface Subject {
     String operation();
